+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Bug fixes on top of `1.6.0`.
   preventing any unwanted intermediary repository garbage collection.
 - The invocation server now correctly reports the total number of
   uncached actions.
+- Allow colons present in remote execution properties.
 
 ### Note for package maintainers